# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "prod - tutorApp",
    "description": "Production url endpoints",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://tutor-app-v1.vercel.app/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/chats/{user2ID}": {
      "get": {
        "tags": ["chats"],
        "summary": "when you now click on one of them, it will take you to a chat page and getChats request would be sent with user2ID in the route param, returning your chat entry with the person, it will be an empty array if there is no chat history",
        "operationId": "getOurChats",
        "parameters": [
          {
            "name": "user2ID",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "example": "645f835a7d15b93c27f16343"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/updateprofile": {
      "patch": {
        "tags": ["user"],
        "summary": "use this to add additional field to user after email confirmation.",
        "operationId": "updateprofile",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "example": {
                      "sex": "female",
                      "subjects": ["maths", "physics"]
                    }
                  }
                ]
              },
              "example": {
                "sex": "female",
                "subjects": ["maths", "physics"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}